refacror: change structure of app
