+++ src/enums/enums.ts
@@ -1,0 +1,17 @@
+export enum HttpCodes {
+  SERVER_ERROR = 500,
+  NOT_FOUND = 404,
+  OK = 200,
+  CREATED = 201,
+  NO_CONTENT = 204,
+}
+
+export enum StatusMsg {
+  SERVER_ERROR_MSG = 'Server error',
+  NOT_FOUND_MSG = 'Not Found',
+  NO_CONTENT_MSG = 'User has been successful deleted',
+}
+
+export enum RequiredError {
+  PARAM_PEQUIRED = 'ID param is required',
+}
