+++ src/server.ts
@@ -3,6 +3,6 @@ import config from './common/config';
 
 const PORT = config.PORT || 4000;
 
-app.listen(4000, () =>
+app.listen(PORT, () =>
   console.log(`App is running on http://localhost:${PORT}`)
 );
