+++ src/types/types.ts
@@ -1,4 +1,4 @@
-import { IBoard, ITask } from './interfaces';
+import { IBoard, ITask } from 'interfaces/interfaces';
 
 export type Board = IBoard | undefined;
